address self-review: the framing counter wrapped, presenting as "healthy"
getFramingErrorCount() narrowed RS485::getErrorCount() -- an unsigned long -- to
uint16_t. That is not academic. The counter bumps once per bad byte-pair, per bad CRC
and per overflow, so 65536 is reachable in a months-long deployment, and a wrap
presents as the value FALLING. The docs tell an operator that a slowly climbing value
means a wiring problem and a flat one means healthy, so a wrap reads as exactly the
reassuring case. Returns uint32_t now; the info row already prints it with %u.

Also corrects two things the doc block asserted:

  * "since boot" -- errorCount_ is zeroed in RS485::begin(), not in the RS485
    constructor's initialiser list, so between init() and setup() it is indeterminate.
    Now says "from begin()" and to not read it earlier.
  * the null guard is channel == NULL, but the value is only meaningful once setup()
    has run. Noted in the comment; returning 0 for an un-set-up socket is the same
    answer either way, so the guard itself is unchanged.

55 checks still pass; avr-g++ -mmcu=atmega328p still clean.

+++ RS485Utils/RS485Utils.cpp
@@ -147,11 +147,14 @@ uint16_t RS485Socket::getRejectCount() {
   return rejectCount;
 }
 
-uint16_t RS485Socket::getFramingErrorCount() {
+uint32_t RS485Socket::getFramingErrorCount() {
   // Null-guarded like packetInProgress()/getLength(): a socket whose channel allocation failed must
-  // report zero rather than dereference.
+  // report zero rather than dereference. Note the value is only MEANINGFUL once setup() has run --
+  // initialized() is the stronger test, but returning 0 for an un-set-up socket is the same answer.
   if (channel == NULL) return 0;
-  return (uint16_t)channel->getErrorCount();
+  // No narrowing: getErrorCount() is an unsigned long, and truncating it would make a wrap look like
+  // the counter going down.
+  return (uint32_t)channel->getErrorCount();
 }
 
 void RS485Socket::setup()
